1.Updated README.md with used technologies 2.Refactor both Home Page PageObject and TestClass names as accordingly RegistrationConfirmationPage.Java and RegistrationConfirmationPageTest.Java
