+++ README.md
@@ -1,1 +1,15 @@
 # Assignment
+
+To run the tests: ' mvn clean install -DspecificTest={TESTNG_FILE_NAME} -Dtestng.dtd.http=true'
+
+Eg: mvn clean install -DspecificTest=testng.xml -Dtestng.dtd.http=true
+
+Used technologies:
+
+Selenium WebDriver
+WebDriver Manager
+Java 20
+Maven
+TestNG
+Allure2 Reporting
+Log4J2 Logging
