--- src/main/java/PageObjects/HomePage.java
@@ -1,14 +1,0 @@
-package PageObjects;
-
-import BaseTest.TestBase;
-
-public class HomePage extends TestBase {
-
-    public HomePage(){
-    }
-
-    //validate Home Page title
-    public String validateHomePageTitle(){
-        return driver.getTitle();
-    }
-}
