+++ src/main/java/PageObjects/RegistrationConfirmationPage.java
@@ -1,0 +1,14 @@
+package PageObjects;
+
+import BaseTest.TestBase;
+
+public class RegistrationConfirmationPage extends TestBase {
+
+    public RegistrationConfirmationPage(){
+    }
+
+   /* //validate Home Page title
+    public String validateHomePageTitle(){
+        return driver.getTitle();
+    }*/
+}
